Factor out lib/web modules.
